# api/openapi.yaml
openapi: 3.1.0
info:
  title: Skawld Maintenance API
  version: 0.6.0-phase5
  description: >
    Phase 5 pilot-ready evidence-first maintenance intelligence contract.
    Learned candidates require exact evidence, human review, deterministic
    evaluation, approval authority, and explicit publication. AI quality
    labels are append-oriented. This API exposes no industrial control action
    or permit authority.
servers:
  - url: http://localhost:8080
    description: Local development
...
paths:
...
  /api/v1/documents:
    get:
      summary: List authorized controlled documents
      tags: [Knowledge]
      operationId: listDocuments
      security: [{sessionCookie: []}, {bearerAuth: []}]
      parameters:
        - {in: query, name: site_id, schema: {type: string, format: uuid}}
        - {in: query, name: cursor, required: false, schema: {type: string}}
        - {in: query, name: page_size, required: false, schema: {type: integer, minimum: 1, maximum: 100, default: 25}}
      responses:
        "200":
          description: Documents and revision lifecycle in authorized scope.
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                required: [items, next_cursor, has_more]
                properties:
                  items: {type: array, items: {type: object}}
                  next_cursor: {type: [string, "null"]}
                  has_more: {type: boolean}
    post:
      summary: Create controlled document metadata
      tags: [Knowledge]
      operationId: createDocument
      security: [{sessionCookie: []}, {bearerAuth: []}]
      parameters: [{$ref: "#/components/parameters/IdempotencyKey"}]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [site_id, document_type, title, authority]
              properties:
                site_id: {type: string, format: uuid}
                document_type:
                  type: string
                  enum: [OEM_MANUAL, SOP, WORK_INSTRUCTION, DATASHEET, INSPECTION_PROCEDURE, SAFETY_PROCEDURE, TECHNICAL_BULLETIN]
                title: {type: string, minLength: 1}
                authority:
                  type: string
                  enum: [OEM, CORPORATE, SITE_APPROVED, REGULATORY, EXPERT_REFERENCE]
                source_reference: {type: string}
      responses:
        "201": {description: Document metadata created.}
        "409": {$ref: "#/components/responses/Problem"}
  /api/v1/documents/{documentID}:
    get:
      summary: Get document and revision validity
      tags: [Knowledge]
      operationId: getDocument
      security: [{sessionCookie: []}, {bearerAuth: []}]
      parameters:
        - {in: path, name: documentID, required: true, schema: {type: string, format: uuid}}
      responses:
        "200": {description: Controlled document with revision history.}
        "404": {$ref: "#/components/responses/Problem"}
  /api/v1/documents/{documentID}/revisions:
    post:
      summary: Create a document revision and applicability
      tags: [Knowledge]
      operationId: createDocumentRevision
      security: [{sessionCookie: []}, {bearerAuth: []}]
      parameters:
        - {in: path, name: documentID, required: true, schema: {type: string, format: uuid}}
        - {$ref: "#/components/parameters/IdempotencyKey"}
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [revision, language, applicability]
              properties:
                revision: {type: string, minLength: 1}
                effective_at: {type: string, format: date-time}
                expires_at: {type: string, format: date-time}
                language: {type: string}
                applicability:
                  type: array
                  minItems: 1
                  items:
                    type: object
                    properties:
                      site_id: {type: string, format: uuid}
                      asset_id: {type: string, format: uuid}
                      asset_class: {type: string}
                      manufacturer: {type: string}
                      model: {type: string}
                      process_service: {type: string}
      responses:
        "201": {description: Draft revision created.}
        "409": {$ref: "#/components/responses/Problem"}
  /api/v1/document-revisions/{revisionID}/ingestion:
    post:
      summary: Atomically queue document extraction and embedding
      tags: [Knowledge]
      operationId: requestDocumentIngestion
      security: [{sessionCookie: []}, {bearerAuth: []}]
      parameters:
        - {in: path, name: revisionID, required: true, schema: {type: string, format: uuid}}
        - {$ref: "#/components/parameters/IdempotencyKey"}
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [attachment_id]
              properties:
                attachment_id: {type: string, format: uuid}
      responses:
        "202": {description: PostgreSQL-backed ingestion job committed with revision state.}
        "422": {$ref: "#/components/responses/Problem"}
  /api/v1/document-revisions/{revisionID}/approve:
    post:
      summary: Approve a ready revision using approval authority
      tags: [Knowledge]
      operationId: approveDocumentRevision
      security: [{sessionCookie: []}, {bearerAuth: []}]
      parameters:
        - {in: path, name: revisionID, required: true, schema: {type: string, format: uuid}}
        - {$ref: "#/components/parameters/IdempotencyKey"}
      requestBody: {$ref: "#/components/requestBodies/ExpectedVersion"}
      responses:
        "200": {description: Revision approved and prior approved revision superseded.}
        "403": {$ref: "#/components/responses/Problem"}
        "409": {$ref: "#/components/responses/Problem"}
  /api/v1/document-revisions/{revisionID}/retire:
    post:
      summary: Retire a document revision
      tags: [Knowledge]
      operationId: retireDocumentRevision
      security: [{sessionCookie: []}, {bearerAuth: []}]
      parameters:
        - {in: path, name: revisionID, required: true, schema: {type: string, format: uuid}}
        - {$ref: "#/components/parameters/IdempotencyKey"}
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [expected_version, reason]
              properties:
                expected_version: {type: integer, format: int64, minimum: 1}
                reason: {type: string, minLength: 1}
      responses:
        "200": {description: Revision retired and immediately ineligible.}
  /api/v1/search:
    post:
      summary: Search an authorization-first eligible evidence set
      description: >
        Tenant, authorized site, approval, validity, supersession, and asset
        applicability are applied in SQL before FTS/vector ranking and RRF.
      tags: [Knowledge]
      operationId: searchEvidence
      security: [{sessionCookie: []}, {bearerAuth: []}]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [site_id, query]
              properties:
                site_id: {type: string, format: uuid}
                asset_id: {type: string, format: uuid}
                query: {type: string, minLength: 1, maxLength: 500}
                limit: {type: integer, minimum: 1, maximum: 20}
                language: {type: string}
                process_service: {type: string}
      responses:
        "200": {description: "Ranked evidence with inspectable lexical, vector, and RRF scores."}
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: skawld_session
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: OIDC access token for native/mobile Authorization Code + PKCE.
  parameters:
    IdempotencyKey:
      in: header
      name: Idempotency-Key
      required: true
      schema:
        type: string
        minLength: 8
        maxLength: 200
...
  requestBodies:
    ExpectedVersion:
      required: true
      content:
        application/json:
          schema:
            type: object
            additionalProperties: false
            required: [expected_version]
            properties:
              expected_version: {type: integer, format: int64, minimum: 1}
  responses:
    Problem:
      description: RFC 9457-style problem response.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
...
  schemas:
...
    Problem:
      type: object
      required: [type, title, status]
      properties:
        type:
          type: string
          format: uri-reference
        title:
          type: string
        status:
          type: integer
          minimum: 400
          maximum: 599
        detail:
          type: string